celery_demo/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
boto==2.49.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.6
django-axes==6.1.1
django-celery-beat==2.5.0
django-ranged-response==0.2.0
django-restframework==0.0.1
django-simple-captcha==0.5.20
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
iniconfig==2.0.0
kombu==5.3.2
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytest==7.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
requests
django-smtp-ssl